Combine the workspace updaters into one file.
