+++ src/client/interpreter/configuration/services/workspaceUpdaterService.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { Uri } from 'vscode';
+import { ConfigurationTarget, Uri } from 'vscode';
 import { IWorkspaceService } from '../../../common/application/types';
 import { IPythonPathUpdaterService } from '../types';
 
@@ -22,5 +22,27 @@ export class WorkspacePythonPathUpdaterService implements IPythonPathUpdaterServ
             pythonPath = path.relative(this.workspace.fsPath, pythonPath);
         }
         await pythonConfig.update('pythonPath', pythonPath, false);
+    }
+}
+
+export class WorkspaceFolderPythonPathUpdaterService implements IPythonPathUpdaterService {
+
+    constructor(
+        private workspaceFolder: Uri,
+        private readonly workspaceService: IWorkspaceService
+    ) { }
+
+    public async updatePythonPath(pythonPath: string): Promise<void> {
+        const pythonConfig = this.workspaceService.getConfiguration('python', this.workspaceFolder);
+        const pythonPathValue = pythonConfig.inspect<string>('pythonPath');
+
+        if (pythonPathValue && pythonPathValue.workspaceFolderValue === pythonPath) {
+            return;
+        }
+        // The workspace folder is guaranteed to be an absolute path.
+        if (pythonPath.startsWith(this.workspaceFolder.fsPath)) {
+            pythonPath = path.relative(this.workspaceFolder.fsPath, pythonPath);
+        }
+        await pythonConfig.update('pythonPath', pythonPath, ConfigurationTarget.WorkspaceFolder);
     }
 }
